Do not fail if no .bowerrc exists in project directory and use Bower's config object instead
The task currently tries to load .bowerrc in the project directory and fails if it's not found. Instead of looking for the file in different places we can just use Bower's config object, which does exactly that already and is also supplemented with the right default values. However, as this plugin uses "bower_components" as the default package folder and Bower has only just made this step in v0.10.0, the dependency version should also be bumped.